requirements.txt
# 核心依赖
chromadb>=0.5.0
anthropic>=0.34.0
sentence-transformers>=3.0.0

# 文档解析
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0

# 中文处理
jieba>=0.42.1

# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# CLI 美化
rich>=13.7.0
typer>=0.12.0

# 工具
python-dotenv>=1.0.0
httpx>=0.27.0
tqdm>=4.66.0
